compute-client: module for durable controller state
This commit moves the part of compute controller state that we want to
make durable in the future, along with any code manipulating it, into a
separate module (`mz_compute_client::controller::durable`).

The motivation for introducing this new module is that the module
boundary allows us to control access to the durable state. Subsequent
commits will lock down the API of `DurableState`, to ensure that every
action that modifies this state is durably recorded.
